% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsifronts.R
\docType{data}
\name{orsifronts}
\alias{orsifronts}
\title{Orsi fronts.}
\format{A \link[sp]{SpatialLinesDataFrame} with variables: \code{name}, \code{front}.}
\usage{
orsifronts
}
\description{
The spatial distributions of the Antarctic Circumpolar Current fronts as presented in
Orsi, A. H., T. Whitworth III and W. D. Nowlin, Jr., 1995: On the meridional extent and fronts of the
Antarctic Circumpolar Current, Deep-Sea Res. I, 42, 641-673.
}
\keyword{datasets}

