% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orsifronts-package.R
\docType{data}
\name{parkfronts}
\alias{parkfronts}
\title{Park/Durand fronts.}
\format{
A Spatial data frame with five line objects, and the identifiers:
\describe{
  \item{name}{long name of the front}
  \item{front}{short name of the front, abbreviation}
}
}
\source{
\doi{10.17882/59800}
}
\usage{
parkfronts
}
\description{
A dataset containing the locations of the Southern Ocean fronts as determined by Park and Durand.
}
\keyword{datasets}
